set REST_AUTH_USER_DETAILS_SERIALIZER, to return object_id of OpenDPUser. Also, moved LoginCard to a separate component.
